// api/types/UpdateAdminRequest.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type UpdateAdminRequest = {
    /**
     * @minLength 1
     * @type string
    */
    fullName: string;
    /**
     * @minLength 1
     * @type string
    */
    email: string;
    /**
     * @minLength 6
     * @maxLength 2147483647
     * @type string | undefined
    */
    password?: string;
    /**
     * @type string | undefined
    */
    department?: string;
    /**
     * @type boolean | undefined
    */
    enabled?: boolean;
    /**
     * @type array | undefined
    */
    permissions?: string[];
};